+++ ingeniamotion/information.py
@@ -1,14 +1,23 @@
-from typing import TYPE_CHECKING, Optional, Tuple
+import os
+from typing import TYPE_CHECKING, Dict, Optional, Tuple
+import xml.etree.ElementTree as ET
 
+from ingenialink.eoe.network import EoENetwork
+from ingenialink.ethernet.network import EthernetNetwork
+from ingenialink.canopen.network import CanopenNetwork
 from ingenialink.register import Register, REG_ACCESS, REG_DTYPE
+import ingenialogger
 
-from ingeniamotion.exceptions import IMRegisterNotExist
+from ingeniamotion.exceptions import IMRegisterNotExist, IMException
 from ingeniamotion.metaclass import MCMetaClass, DEFAULT_AXIS, DEFAULT_SERVO
+from ingeniamotion.enums import COMMUNICATION_TYPE
 from ingeniamotion.comkit import create_comkit_dictionary
 
 
 if TYPE_CHECKING:
     from ingeniamotion.motion_controller import MotionController
+
+logger = ingenialogger.get_logger(__name__)
 
 
 class Information(metaclass=MCMetaClass):
@@ -154,3 +163,186 @@ class Information(metaclass=MCMetaClass):
 
         """
         return create_comkit_dictionary(coco_dict_path, moco_dict_path, dest_file)
+
+    def get_product_name(self, alias: str = DEFAULT_SERVO) -> Optional[str]:
+        """Get the product name of the drive.
+
+        Args:
+            alias: alias of the servo.
+
+        Returns:
+            If it exists for example: "EVE-NET-E", "CAP-NET-E", etc.
+        """
+        drive = self.mc.servos[alias]
+        product_name = drive.dictionary.part_number
+        if not product_name is None:
+            return f"{product_name}"
+        return None
+
+    def get_node_id(self, alias: str = DEFAULT_SERVO) -> int:
+        """Get the node ID for CANopen communications.
+
+        Args:
+            alias: alias of the servo.
+
+        Returns:
+            Node ID of the drive.
+        """
+        net = self.mc._get_network(alias)
+        drive = self.mc.servos[alias]
+        if isinstance(net, CanopenNetwork):
+            return int(drive.target)
+        else:
+            raise IMException("You need a CANopen communication to use this function")
+
+    def get_ip(self, alias: str = DEFAULT_SERVO) -> str:
+        """Get the IP for Ethernet communications.
+
+        Args:
+            alias: alias of the servo.
+
+        Returns:
+            IP of the drive.
+        """
+        net = self.mc._get_network(alias)
+        drive = self.mc.servos[alias]
+        if isinstance(net, EthernetNetwork):
+            return str(drive.target)
+        else:
+            raise IMException("You need an Ethernet communication to use this function")
+
+    def get_slave_id(self, alias: str = DEFAULT_SERVO) -> int:
+        """Get the slave ID for EoE communications.
+
+        Args:
+            alias: alias of the servo.
+
+        Returns:
+            Slave ID of the drive.
+        """
+        net = self.mc._get_network(alias)
+        drive = self.mc.servos[alias]
+        if isinstance(net, EoENetwork):
+            return int(net._configured_slaves[drive.target])
+        else:
+            raise IMException("You need a CANopen communication to use this function")
+
+    def get_name(self, alias: str = DEFAULT_SERVO) -> str:
+        """Get the drive's name.
+
+        Args:
+            alias: Alias of the servo.
+        Returns:
+            The name of the drive.
+        """
+        drive = self.mc.servos[alias]
+        drive_name = drive.name
+        return f"{drive_name}"
+
+    def get_communication_type(self, alias: str = DEFAULT_SERVO) -> COMMUNICATION_TYPE:
+        """Get the connected drive's communication type.
+
+        Args:
+            alias: alias of the connected drive.
+
+        Returns:
+            CANopen, Ethernet, or EtherCAT.
+
+        """
+        drive_network = self.mc._get_network(alias)
+        if isinstance(drive_network, CanopenNetwork):
+            communication_type = COMMUNICATION_TYPE.Canopen
+        elif isinstance(drive_network, EthernetNetwork):
+            communication_type = COMMUNICATION_TYPE.Ethernet
+        elif isinstance(drive_network, EoENetwork):
+            communication_type = COMMUNICATION_TYPE.Ethercat
+        return communication_type
+
+    def get_full_name(self, alias: str = DEFAULT_SERVO) -> str:
+        """Return the full name of the drive [Product name] [Name] ([Target]).
+
+        Args:
+            alias: Drive alias.
+
+        Returns:
+            Full name.
+        """
+        prod_name = self.get_product_name(alias)
+        name = self.get_name(alias)
+        full_name = f"{prod_name} - {name}"
+        net = self.mc._get_network(alias)
+        if isinstance(net, EthernetNetwork):
+            ip = self.get_ip(alias)
+            full_name = f"{full_name} ({ip})"
+        return full_name
+
+    def get_subnodes(self, alias: str = DEFAULT_SERVO) -> int:
+        """Return the number of subnodes.
+
+        Args:
+            alias: Drive alias.
+
+        Returns:
+            Number of subnodes.
+        """
+        drive = self.mc.servos[alias]
+        return int(drive.subnodes)
+
+    def get_categories(self, alias: str) -> Dict[str, str]:
+        """Return dictionary categories instance.
+
+        Args:
+            alias: Drive alias.
+
+        Returns:
+            Categories instance.
+        """
+        drive = self.mc.servos[alias]
+        dictionary_categories = drive.dictionary.categories
+        category_ids = dictionary_categories.category_ids
+        categories: Dict[str, str] = {}
+        for cat_id in category_ids:
+            categories[cat_id] = dictionary_categories.labels(cat_id)["en_US"]
+        return categories
+
+    def get_dictionary_file_name(self, alias: str) -> str:
+        """Return dictionary file name.
+
+        Args:
+            alias: Drive alias.
+
+        Returns:
+            Dictionary file name.
+        """
+        drive = self.mc.servos[alias]
+        return str(os.path.basename(drive.dictionary.path))
+
+    # TODO: INGM-333 - Once ingenialink has the encoded image
+    def get_encoded_image_from_dictionary(self, alias: str) -> Optional[str]:
+        """Get the encoded product image from a drive dictionary.
+        This function reads a dictionary of a drive, and it parses whether the dictionary file has a
+        DriveImage tag and its content.
+        Args:
+            alias: Alias of the drive.
+        Returns:
+            The encoded image or NoneType object.
+        """
+        drive = self.mc.servos[alias]
+        # Read encoded image in XDF dictionary file
+        dictionary_path = drive.dictionary.path
+        try:
+            with open(dictionary_path, "r", encoding="utf-8") as xdf_file:
+                tree = ET.parse(xdf_file)
+        except FileNotFoundError:
+            raise FileNotFoundError(f"There is not any xdf file in the path: {dictionary_path}")
+        root = tree.getroot()
+        try:
+            image_element = root.findall(f"./DriveImage")
+            if image_element[0].text is not None and image_element[0].text.strip():
+                return f"{image_element[0].text}"
+            else:
+                # If the content in DriveImage tag is empty
+                return None
+        except IndexError:
+            # If there is no DriveImage tag in dictionary file
+            return None