+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Added
 - Functions needed to load firmware to a Motion Core (MoCo).
 - COM-KIT support.
+- In Information class: get_product_name, get_node_id, get_ip, get_slave_id, get_name, get_communication_type, get_full_name, get_subnodes, get_categories, get_dictionary_file_name.
+- In Configuation class: get_drive_info_coco_moco, get_product_code, get_revision_number, get_serial_number, get_fw_version
+- no_connection marker is added for unit testing.
 - Add mypy into the project.
 - Add types for all functions and classes.
 - Make and pass the first static type analysis.